Removed struve and modstruve
